+++ zetta_utils/mazepa/flows.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import functools
 from contextlib import contextmanager
 from typing import (
     Any,
@@ -20,7 +21,7 @@ from typing import (
 import attrs
 from typing_extensions import ParamSpec
 
-from . import id_generators
+from . import id_generation
 from .dependency import Dependency
 from .task_execution_env import TaskExecutionEnv
 from .tasks import Task
@@ -137,7 +138,9 @@ class _FlowSchema(Generic[P]):
     """
 
     fn: Callable[P, FlowFnReturnType]
-    id_fn: Callable[[Callable, dict], str] = attrs.field(default=id_generators.get_unique_id)
+    id_fn: Callable[[Callable, list, dict], str] = attrs.field(
+        default=functools.partial(id_generation.generate_invocation_id, prefix="flow")
+    )
     task_execution_env: TaskExecutionEnv = attrs.field(factory=TaskExecutionEnv)
 
     def __call__(
@@ -145,7 +148,7 @@ class _FlowSchema(Generic[P]):
         *args: P.args,
         **kwargs: P.kwargs,
     ) -> Flow:
-        id_ = self.id_fn(self.fn, kwargs)
+        id_ = self.id_fn(self.fn, list(args), kwargs)
         result = _Flow(
             fn=self.fn,
             id_=id_,
